requirements.txt
lightkurve
numpy
matplotlib
scipy
astropy
transitleastsquares
setuptools
opencv-python
Pillow
